# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.oas36ty.test:8000/v1",
            "description": "Localhost"
        }
    ],
    "paths": {
...
        "/tasks": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Tasks",
                "description": "Tasks",
                "operationId": "getTasks",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/tenant--header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success Message!"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "lead"
                                                    },
                                                    "subject": {
                                                        "type": "string",
                                                        "example": "Task subject"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Task description"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error Message!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "tasks"
                ],
                "summary": "Create Task",
                "description": "Create Task",
                "operationId": "postTask",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/tenant--header"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "description": "",
                                        "type": "string",
                                        "example": "lead"
                                    },
                                    "subject": {
                                        "description": "",
                                        "type": "string",
                                        "example": "Task subject"
                                    },
                                    "description": {
                                        "description": "",
                                        "type": "string",
                                        "example": "Task description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success Message!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error Message!"
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "INVALID"
                                        },
                                        "errors": {
                                            "properties": {
                                                "subject": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The selected subject is invalid."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "parameters": {
            "tenant--header": {
                "name": "X-Tenant",
                "in": "header",
                "description": "Tenant ID",
                "schema": {
                    "type": "string",
                    "default": "oas36ty"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Execute Login or Signup APIs to get the authentication token",
                "name": "Token Based Authentication",
                "scheme": "bearer"
            }
        }
    }
}